third_party/WebKit/LayoutTests/editing/deleting/delete-line-003.html

<!doctype HTML>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
  '<div contenteditable style="font-size: 24px; width: 600px;">There is a tide^<br>in the affairs of men|<br>Which taken at the flood</div>',
  'delete',
  '<div contenteditable style="font-size: 24px; width: 600px;">There is a tide|<br>Which taken at the flood</div>'),
  'Editing test');
</script>
